% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phaser.R
\name{bi_allelic_phaser}
\alias{bi_allelic_phaser}
\title{Phase mixed alleles using the 'bi_allelic' method}
\usage{
bi_allelic_phaser(mad, ncores)
}
\arguments{
\item{mad}{A \code{data.frame} with three columns that represents the indices of
the mixed alleles, the indices of the SNPs they are found on, and their
allelic depths supporting both reads.}

\item{ncores}{A numeric that represents the number of cores to be used during
the phasing process. Default is 4.}
}
\description{
Phase mixed alleles using the 'bi_allelic' method
}
\keyword{internal}
