% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pop_genetics.R
\name{calculate_IBS}
\alias{calculate_IBS}
\title{Generate dissimilarity matrix (1-IBS) between all pairs of isolates}
\usage{
calculate_IBS(snpdata, mat_name = "GT")
}
\arguments{
\item{snpdata}{SNPdata object}

\item{mat_name}{the name of the genotype table to be used. default="GT"}
}
\value{
a \code{SNPdata} object with an extra field: \code{IBS}
}
\description{
Generate dissimilarity matrix (1-IBS) between all pairs of isolates
}
\examples{
\dontrun{
  snpdata <- calculate_IBS(snpdata, mat_name = "GT")
 }

}
