% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pop_genetics.R
\name{calculate_iR}
\alias{calculate_iR}
\title{Calculate iR index to detect loci with excess of IBD sharing}
\usage{
calculate_iR(
  snpdata,
  mat_name = "Phased",
  family = "Location",
  number_cores = 4L
)
}
\arguments{
\item{snpdata}{SNPdata object}

\item{mat_name}{the name of the genotype table to be used. default = "Phased"}

\item{family}{the name of the column, in the metadata table, to be used to
represent the sample's population}

\item{number_cores}{the number of cores to be used. default = 4}
}
\value{
a \code{SNPdata} object with an extra field: \code{iR}
}
\description{
Calculate iR index to detect loci with excess of IBD sharing
}
\examples{
\dontrun{
  snpdata <- calculate_iR(
    snpdata,
    mat_name     = "Phased",
    family       = "Location",
    number_cores = 4
 )
 }
}
