% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pop_genetics.R
\name{calculate_LD}
\alias{calculate_LD}
\title{Calculate LD R^2 between pairs of loci}
\usage{
snpdata = calculate_LD(snpdata, min.r2=0.2, inter.chrom=FALSE, chroms=c("Pf3D7_04_v3","Pf3D7_05_v3"))
}
\arguments{
\item{snpdata}{SNPdata object}

\item{min.r2}{the minimum r2 value below which the LD value is not reported}

\item{inter.chrom}{whether to calculate inter-chromosomal LD. FALSE by default}

\item{chroms}{a vector of chromosome names for which LD should be calculated}
}
\value{
SNPdata object with an extra field: LD
}
\description{
Calculate LD R^2 between pairs of loci
}
\details{
the output file from LD calculation could be large. In order to reduce the size of that file, it's recommended to specify the list of chromosomes for which LD should be calculated using the \code{chroms} option
}
