% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pop_genetics.R
\name{calculate_relatedness}
\alias{calculate_relatedness}
\title{Estimate the relatedness between pairs of isolates}
\usage{
calculate_relatedness(
  snpdata,
  mat_name = "Imputed",
  from = "Location",
  sweepRegions = NULL,
  groups = NULL
)
}
\arguments{
\item{snpdata}{SNPdata object}

\item{mat_name}{the name of the genotype table to be used. default="GT"}

\item{from}{the name of the column, in the metadata table, to be used to represent the sample's population}

\item{sweepRegions}{a data frame with the genomic coordinates of the regions of the genome to be discarded. This should contain the following 3 columns:
\enumerate{
\item Chrom: the chromosome ID
\item Start: the start position of the region on the chromosome
\item End: the end position of the region on the chromosome
}}

\item{groups}{a vector of character. If specified, relatedness will be generated between these groups}
}
\value{
SNPdata object with an extra field: relatedness. This will contain the relatedness data frame of 3 columns and its correspondent matrix
}
\description{
Estimate the relatedness between pairs of isolates
}
\details{
The relatedness calculation is based on the model developed by Aimee R. Taylor and co-authors. https://journals.plos.org/plosgenetics/article?id=10.1371/journal.pgen.1009101
}
\examples{
\dontrun{
  calculate_relatedness(snpdata, mat_name="GT", family="Location", 
  sweepRegions=NULL, groups=c("Chogen","DongoroBa"))
 }
}
