% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pop_genetics.R
\name{calculate_wcFst}
\alias{calculate_wcFst}
\title{Calculate Weir & Cockerham's Fst}
\usage{
calculate_wcFst(snpdata, from, groups)
}
\arguments{
\item{snpdata}{\code{SNPdata} object}

\item{from}{the metadata column that contains the sample's population
information.}

\item{groups}{a vector of population names. Every sample in the metadata file
is associated to its population of origin. The differentiation index will
be estimated between these groups.}
}
\value{
a \code{SNPdata} object with an extra field named as \strong{Fst}. This is a
\code{list} of data frames with the Fst values for each pair of comparison.
Every data frame contains N rows (where N is the number of loci) and the
following 7 columns:
\enumerate{
\item the chromosome ID
\item the SNPs positions
\item the allele frequency in the first group
\item the allele frequency in the second group
\item the resulting Fst values
\item the p-values associated with the Fst results
\item the p-values corrected for multiple testing using the
Benjamini-Hochberg method
}
}
\description{
Calculate Weir & Cockerham's Fst
}
\examples{
\dontrun{
  snpdata <- calculate_wcFst(
    snpdata,
    groups = c("Senegal", "Gambia"),
    from   = "Country"
  )
}
}
