% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_genotype_matrix}
\alias{check_genotype_matrix}
\title{Check whether the specified genotype matrix is part of the \code{snpdata} object}
\usage{
check_genotype_matrix(snpdata, genotype)
}
\arguments{
\item{snpdata}{a \code{SNPdata} object}

\item{genotype}{The name of the genotype matrix from which the mixed
genotypes will be phased. This can be either the raw genotype matrix
(\code{GT}) or the imputed genotype matrix (\code{Imputed}) or any name given to
the target genotype matrix.}
}
\value{
Invisibly returns TRUE if the specified genotype matrix is part of
the \code{snpdata} object, raises an error otherwise.
}
\description{
Check whether the specified genotype matrix is part of the \code{snpdata} object
}
\keyword{internal}
