% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_maf.R
\name{compute_maf}
\alias{compute_maf}
\title{Calculate minor allele frequency (MAF) at every loci}
\usage{
compute_maf(snpdata, genotype = "GT", include_het = FALSE, name = NULL)
}
\arguments{
\item{snpdata}{An object of class \code{SNPdata}}

\item{genotype}{A string with the name of the matrix to be used. Default is
"GT". The other possible values are "Phased", "Imputed", "Phased_imputed".}

\item{include_het}{A Boolean that specifies whether to account for the
heterozygous allele or not. This can only be activated when
\code{genotype = "GT"} or \code{genotype = "Imputed"}.}

\item{name}{A character with the name of the new column that will created to
store the MAF values.}
}
\value{
The input \code{SNPdata} object with following 2 additional columns in the
\strong{details} table:
\enumerate{
\item MAF: minor allele frequency at every each SNPs
\item MAF_allele: the code for the minor allele. Possible values are:
\strong{1}: the alternate allele is the minor allele,
\strong{0}: the reference allele is the minor allele,
\strong{0/1}: the heterozygous allele is the minor allele,
\strong{0=1}: the reference and alternate alleles have the same frequencies,
\strong{0=1=2}: the three alleles have the same frequencies.
}
}
\description{
Calculate minor allele frequency (MAF) at every loci
}
\details{
If \code{include_het = FALSE}, the mixed alleles will not be considered
in the MAF calculation.
}
