% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pop_genetics.R
\name{compute_rhat_hmm}
\alias{compute_rhat_hmm}
\title{Title}
\usage{
compute_rhat_hmm(freqquencies, distances, ys, epsilon)
}
\arguments{
\item{freqquencies}{}

\item{distances}{}

\item{ys}{}

\item{epsilon}{}
}
\description{
Title
}
