% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pop_genetics.R
\name{construct_genotype_file}
\alias{construct_genotype_file}
\title{Title}
\usage{
construct_genotype_file(pops, details, mat, metadata, from, sweep_regions)
}
\arguments{
\item{pops}{}

\item{details}{}

\item{mat}{}

\item{metadata}{}

\item{from}{}

\item{sweep_regions}{}
}
\description{
Title
}
