% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pop_genetics.R
\name{create_relatedness_matrix}
\alias{create_relatedness_matrix}
\title{Title}
\usage{
create_relatedness_matrix(ibd, metadata, pops, from)
}
\arguments{
\item{ibd}{}

\item{metadata}{}

\item{pops}{}

\item{from}{}
}
\description{
Title
}
