% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drop_snps.R
\name{drop_snps}
\alias{drop_snps}
\title{Drop a of set SNPs from a \code{SNPdata} object}
\usage{
drop_snps(
  snpdata,
  snp_to_be_dropped = NULL,
  chrom = NULL,
  start = NULL,
  end = NULL
)
}
\arguments{
\item{snpdata}{The input \code{SNPdata} object}

\item{snp_to_be_dropped}{A data frame with 2 columns: "Chrom" and "Pos".}

\item{chrom}{The chromosome from which loci should be dropped}

\item{start}{The starting position of the region to be discarded}

\item{end}{The end position of the region to be discarded}
}
\value{
A \code{SNPdata} object where the specified SNPs have been removed.
}
\description{
Drop a of set SNPs from a \code{SNPdata} object
}
\details{
When 'snp_to_be_dropped' is not NULL (i.e. the genomic
coordinates of snps to be removed are in a data frame), then the rest of
the arguments can be ignored or set to NULL (chrom = NULL,
start = NULL, end = NULL)
}
