% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access_genotype_data.R
\name{extract_allelic_depth}
\alias{extract_allelic_depth}
\title{Extract the allelic depth from a variant call file (VCF)}
\usage{
extract_allelic_depth(file)
}
\arguments{
\item{file}{A character with the path to the input VCF file.}
}
\value{
A data frame with the allelic depth supporting each of the two
alleles at any given locus across all samples.
}
\description{
Extract the allelic depth from a variant call file (VCF)
}
\examples{
\dontrun{
  allelic_depth <- extract_allelic_depth(
    file = system.file("extdata", "Input_Data.vcf.gz", package = "mpbr")
  )
}

}
