% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access_genotype_data.R
\name{extract_genotype}
\alias{extract_genotype}
\title{Extract the genotypes from a variant call file (VCF)}
\usage{
extract_genotype(file)
}
\arguments{
\item{file}{A character with the path to the input VCF file.}
}
\value{
A data frame with the genotype data together with the SNPs genomic
coordinates and their associated calling quality
}
\description{
Extract the genotypes from a variant call file (VCF)
}
\examples{
\dontrun{
  extract_genotype <- extract_allelic_depth(
    file = system.file("extdata", "Input_Data.vcf.gz", package = "mpbr")
  )
}

}
\author{
Banky
}
