% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_snps_and_samples.R
\name{filter_hard}
\alias{filter_hard}
\title{Hard filter on SNPs and samples}
\usage{
filter_hard(
  snpdata,
  min_qual,
  maf_cutoff,
  max_missing_samples,
  max_missing_sites
)
}
\arguments{
\item{snpdata}{An object of class \code{SNPdata}}

\item{min_qual}{A numeric value representing the minimum call quality score
below which a loci will be discarded. Default is 1000.}

\item{maf_cutoff}{A numeric value between \code{0} and \code{1} representing the MAF
cut-off. Loci with \code{MAF < maf_cutoff} will be discarded.
Default is 0.01.}

\item{max_missing_samples}{A numeric value between \code{0} and \code{1} representing
the maximum fraction of missing samples above which a loci should be
discarded. Default is 0.2.}

\item{max_missing_sites}{A numeric value between \code{0} and \code{1} representing the
maximum fraction of missing sites above which a sample should be
discarded. Default is 0.2.}
}
\value{
A subset of the input \code{SNPdata} object where samples and SNPs
that do not meet the defined thresholds have been remove.
}
\description{
Hard filter on SNPs and samples
}
\keyword{internal}
