% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_snps_and_samples.R
\name{filter_soft}
\alias{filter_soft}
\title{Soft filter on SNPs and samples}
\usage{
filter_soft(snpdata, missingness, maf)
}
\arguments{
\item{snpdata}{An object of class \code{SNPdata}}

\item{missingness}{A numeric vector of three elements that specifies the
cut-off range to use for the soft filtration. When provided, SNPs and
samples will be filtered across this range. Default is
\code{missingness = c(from = 0.1, to = 0.9, by = 0.1)}.}

\item{maf}{A numeric vector of three elements that specifies the
cut-offs to use for the soft filtration on MAF. When provided, SNPs will
be filtered across this range. Default is
\code{maf = c(from = 0.01, to = 0.5, by = 0.01)}.}
}
\value{
A list with the following two elements of type data frame:
\describe{
\item{\code{missingness}}{A data frame with three columns showing the remaining
proportion of samples and loci after every missingness cut-off.}
\item{\code{maf}}{A data frame with two columns representing the remaining
proportion of loci at every MAF cut-off}.
}
}
\description{
Soft filter on SNPs and samples
}
\keyword{internal}
