% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation.R
\name{gene_annotation}
\alias{gene_annotation}
\title{Associate each SNPs to a gene id, name and description based on the gene
they fall under}
\usage{
gene_annotation(target_gtf, genomic_coordinates)
}
\arguments{
\item{target_gtf}{A data frame with the gene annotation}

\item{genomic_coordinates}{A data frame with the SNPs genomic coordinates}
}
\value{
A vector with gene annotation. This should be of the same length as
the number of SNPs.
}
\description{
Associate each SNPs to a gene id, name and description based on the gene
they fall under
}
\keyword{internal}
