% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation.R
\name{get_clean_name}
\alias{get_clean_name}
\title{Clean gene names}
\usage{
get_clean_name(y)
}
\arguments{
\item{y}{The gene name}
}
\value{
A string with the cleaned gene name
}
\description{
Clean gene names
}
\keyword{internal}
