% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation.R
\name{get_gene_annotation}
\alias{get_gene_annotation}
\title{Add gene ontology and names annotation details to every SNPs in the table
that contains their genomic coordinates}
\usage{
get_gene_annotation(genomic_coordinates, go, bed, nthreads = 4L)
}
\arguments{
\item{genomic_coordinates}{The table with SNPs genomic coordinates}

\item{go}{A data frame with the gene ontology annotation details}

\item{bed}{A data frame with the gene name annotation}

\item{nthreads}{The number of cores to be used}
}
\value{
An object of type \code{data.frame} with the SNPs genomic coordinates and
their corresponding annotation details
}
\description{
Add gene ontology and names annotation details to every SNPs in the table
that contains their genomic coordinates
}
\keyword{internal}
