% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pop_genetics.R
\name{get_genotypes}
\alias{get_genotypes}
\title{Title}
\usage{
get_genotypes(
  ped_map,
  reference_ped_map = NULL,
  maf = 0.01,
  isolate_max_missing = 0.1,
  snp_max_missing = 0.1,
  chromosomes = NULL,
  input_map_distance = "cM",
  reference_map_distance = "cM"
)
}
\arguments{
\item{reference_ped_map}{}

\item{maf}{}

\item{isolate_max_missing}{}

\item{snp_max_missing}{}

\item{chromosomes}{}

\item{input_map_distance}{}

\item{reference_map_distance}{}
}
\description{
Title
}
