% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_maf.R
\name{get_maf}
\alias{get_maf}
\title{Get the minor allele frequency (MAF) and the corresponding allele}
\usage{
get_maf(mat)
}
\arguments{
\item{mat}{a matrix with the 2 or 3 columns. Every column should contain the
the count of the allele of interest across all samples.}
}
\value{
a \code{vector} of the MAF and the corresponding allele
}
\description{
Get the minor allele frequency (MAF) and the corresponding allele
}
\keyword{internal}
