% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_region_variants.R
\name{get_region_variants}
\alias{get_region_variants}
\title{Check whether a VCF file contains SNPs overlapping a set of genes of interest}
\usage{
get_region_variants(snpdata, regions)
}
\arguments{
\item{snpdata}{A \if{html}{\out{<SNPdata>}} object obtained from \code{get_snpdata()} function}

\item{regions}{A character with the path to the \code{BED} file with the
genomic coordinates of the regions of interest. It must contain at least
the CHROM, START, and END columns}
}
\value{
A list of two elements: the variants overlapping each region and the
region's variant counts.
}
\description{
Check whether a VCF file contains SNPs overlapping a set of genes of interest
}
\examples{
\dontrun{
  res <- get_region_variants(
    snpdata = snpdata,
    regions = file.path("data", "genes_cordinates_on_v68.txt")
  )
  region_variants <- res[["region_variants"]]
  variant_count <- res[["variant_count"]]
}
}
