% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_sample_ids}
\alias{get_sample_ids}
\title{Extract the sample ids from a VCF file}
\usage{
get_sample_ids(vcf_file)
}
\arguments{
\item{vcf_file}{The path to the input VCF file}
}
\value{
A vector with the sample ids
}
\description{
Extract the sample ids from a VCF file
}
\keyword{internal}
