% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phaser.R
\name{least_frequent_phaser}
\alias{least_frequent_phaser}
\title{Phase mixed alleles using the 'least_frequent' method}
\usage{
least_frequent_phaser(mad, ncores, details)
}
\arguments{
\item{mad}{A \code{data.frame} with three columns that represents the indices of
the mixed alleles, the indices of the SNPs they are found on, and their
allelic depths supporting both reads.}

\item{ncores}{A numeric that represents the number of cores to be used during
the phasing process. Default is 4.}

\item{details}{A data frame with the SNPs genomic coordinates and their
annotations.}
}
\description{
Phase mixed alleles using the 'least_frequent' method
}
\keyword{internal}
