% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pop_genetics.R
\name{loglikelihood}
\alias{loglikelihood}
\title{compute loglikelihood}
\usage{
loglikelihood(k, r, ys, f, gendist, epsilon, rho = 7.4 * 10L^(-7L))
}
\arguments{
\item{k}{k}

\item{r}{r}

\item{ys}{ys}

\item{f}{f}

\item{gendist}{gendist}

\item{epsilon}{epsilon}

\item{rho}{rho}
}
\value{
test
}
\description{
compute loglikelihood
}
