% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phaser.R
\name{major_call_phaser}
\alias{major_call_phaser}
\title{Phase mixed alleles using the 'major_call' method}
\usage{
major_call_phaser(mad, ncores = 4)
}
\arguments{
\item{mad}{A \code{data.frame} with three columns that represents the indices of
the mixed alleles, the indices of the SNPs they are found on, and their
allelic depths supporting both reads.}

\item{ncores}{A numeric that represents the number of cores to be used during
the phasing process. Default is 4.}
}
\description{
Phase mixed alleles using the 'major_call' method
}
\keyword{internal}
