% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phase_mixed_genotypes.R
\name{phase}
\alias{phase}
\title{Phase mixed genotypes}
\usage{
phase(
  snpdata,
  genotype = "GT",
  heterozygous = FALSE,
  min_ad = 2,
  ratio = 0.1,
  ncores = 4,
  phase_equal_ad = "bernoulli"
)
}
\arguments{
\item{snpdata}{a \code{SNPdata} object}

\item{genotype}{The name of the genotype matrix from which the mixed
genotypes will be phased. This can be either the raw genotype matrix
(\code{GT}) or the imputed genotype matrix (\code{Imputed}) or any name given to
the target genotype matrix.}

\item{heterozygous}{A logical that is used to determine whether keep \code{true}
heterozygous sites (recode them as \code{2}) or not. Default is \code{FALSE}. A site
is considered as a true heterozygous when the following two conditions are
met:
\enumerate{
\item condition 1: the \code{ratio} in allelic depth > the specified ratio
cut-off.
\item condition2: the \code{min_ad} >= the specified minimum allelic depth i.e.
(ratio > 0.1 & min_ad >= 2 by default).
}}

\item{min_ad}{An integer that represents one of the criteria described below.
The number of reads supporting the reference or alternative alleles must
be greater than or equal to this cut-off. Default is \code{2}.}

\item{ncores}{A numeric that represents the number of cores to be used during
the phasing process. Default is 4.}

\item{phase_equal_ad}{A character with the name of the phasing approach
used when the allelic depths of the two alleles are equal. The possible
values are:
\enumerate{
\item bernoulli: the mixed allele is recoded into \code{0} or \code{1} based on a
Bernoulli simulation with a parameter equal to the minor allele
frequency of the SNPs on which the site is found.
\item minor_allele: the mixed allele is transformed into the least
represented allele (i.e. the allele with the minor frequency in the
population).
}}

\item{ration}{A float that represents one of the criteria described below.
The ratio between the number of reads supporting the two alleles must be
greater than or equal to this cut-off. Default is \code{0.1}.}
}
\value{
a \code{SNPdata} object with an additional table named as \strong{Phased} that
contains the phased genotype data.
}
\description{
Mixed genotype phasing is performed based on the number of reads supporting
each allele at a heterozygous site (allelic depth). A mixed allele will be
recoded into either \code{0} or \code{1} depending on which one has a highest allelic
depth or into \code{2} when \code{heterozygous} is set to \code{TRUE} i.e. to consider
mixed alleles as a third allele.
}
\examples{
# get SNPdata object
snpdata <- get_snpdata(
  vcf_file = system.file("extdata", "Input_Data.vcf.gz", package = "mpbr"),
  meta_file = system.file("extdata", "SampleMetadata.RDS", package = "mpbr"),
  output_dir = tempdir()
)

# perform mixed genotypes phasing
snpdata <- phase(
  snpdata = snpdata,
  genotype = "GT",
  heterozygous = FALSE, # recode all mixed alleles as 0 or 1
  ncores = 4,
  phase_equal_ad = "bernoulli"
)
 
}
