% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_manip.R
\name{phase_data}
\alias{phase_data}
\title{Phase the mixed genotypes}
\usage{
phase_data(genotype, depth)
}
\arguments{
\item{genotype}{a vector of genotype data}

\item{depth}{a vector of allelic depth}
}
\value{
a vector of phased genotypes
}
\description{
Phase the mixed genotypes
}
\keyword{internal}
