% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phase_mixed_genotypes.R
\name{phase_missing_ref_or_alt}
\alias{phase_missing_ref_or_alt}
\title{Phase genotype when either both reference or alternate allelic depth or one
of them is missing.}
\usage{
phase_missing_ref_or_alt(genotype)
}
\arguments{
\item{genotype}{a vector of integers}
}
\description{
Phase genotype when either both reference or alternate allelic depth or one
of them is missing.
}
\keyword{internal}
