% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phase_mixed_genotypes.R
\name{phase_one_allele}
\alias{phase_one_allele}
\title{Phase the mixed genotypes for one SNP across all samples}
\usage{
phase_one_allele(genotype, depth)
}
\arguments{
\item{genotype}{a vector of genotype data}

\item{depth}{a vector of allelic depth}
}
\value{
a vector of phased genotypes
}
\description{
Phase the mixed genotypes for one SNP across all samples
}
\keyword{internal}
