% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phase_mixed_genotypes.R
\name{phase_ref_alt_0}
\alias{phase_ref_alt_0}
\title{Phase genotype when reference and alternate allele counts = 0}
\usage{
phase_ref_alt_0(ref_count, alt_count)
}
\arguments{
\item{ref_count}{the reference allele count}

\item{alt_count}{the alternate allele count}
}
\description{
Phase genotype when reference and alternate allele counts = 0
}
\keyword{internal}
