% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phase_mixed_genotypes.R
\name{phase_ref_or_alt_0}
\alias{phase_ref_or_alt_0}
\title{Phase genotype when either the reference or alternate allele counts = 0}
\usage{
phase_ref_or_alt_0(genotype, ref, alt)
}
\arguments{
\item{genotype}{a vector of integers}

\item{ref}{the population reference allele count on a specific locus}

\item{alt}{the population alternate allele count on a specific locus}
}
\description{
Phase genotype when either the reference or alternate allele counts = 0
}
\keyword{internal}
