% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phaser.R
\name{phaser}
\alias{phaser}
\title{Perform phasing of a given mixed genotype}
\usage{
phaser(
  mad,
  method,
  ratio,
  min_ad,
  alternative_method,
  ncores = 4,
  details = NULL
)
}
\arguments{
\item{mad}{A \code{data.frame} with three columns that represents the indices of
the mixed alleles, the indices of the SNPs they are found on, and their
allelic depths supporting both reads.}

\item{min_ad}{An integer that represents one of the criteria described below.
The number of reads supporting the reference or alternative alleles must
be greater than or equal to this cut-off. Default is \code{2}.}

\item{ncores}{A numeric that represents the number of cores to be used during
the phasing process. Default is 4.}

\item{details}{A data frame with the SNPs genomic coordinates and their
annotations.}
}
\value{
The input data frame with an additional column named as
\code{phased_genotype}.
}
\description{
Perform phasing of a given mixed genotype
}
\keyword{internal}
