% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_snps_and_samples.R
\name{plot_distribution}
\alias{plot_distribution}
\title{Plot the SNPs distribution across the genome}
\usage{
plot_distribution(snpdata, window_size = 100)
}
\arguments{
\item{snpdata}{An object of class \code{SNPdata}}

\item{window_size}{An integer that is used to specify the window size to
consider. The chromosomes will be split by this window size and the
number of SNPs per window is evaluated then plotted.}
}
\value{
Displays the distribution of the SNPs across the genome.
}
\description{
Plot the SNPs distribution across the genome
}
