% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printing.R
\name{print.SNPdata}
\alias{print.SNPdata}
\title{Print the \code{SNPdata} object}
\usage{
\method{print}{SNPdata}(x, ...)
}
\arguments{
\item{x}{An object of class \code{SNPdata}}

\item{...}{further arguments passed to or from other methods.}
}
\value{
prints the structure of the \code{SNPdata} object
}
\description{
Print the \code{SNPdata} object
}
