% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drop_snps.R
\name{remove_region_from_snpdata}
\alias{remove_region_from_snpdata}
\title{Remove SNPs with a specified region of the genome.}
\usage{
remove_region_from_snpdata(snpdata, chrom, start, end)
}
\description{
Remove SNPs with a specified region of the genome.
}
\keyword{internal}
