% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drop_samples.R
\name{remove_samples}
\alias{remove_samples}
\title{Remove a set of samples from a \if{html}{\out{<SNPdata>}} object}
\usage{
remove_samples(snpdata, samples)
}
\arguments{
\item{snpdata}{The input \if{html}{\out{<SNPdata>}} object}

\item{samples}{A vector or a file with one column containing the list of
samples to be dropped from the provided \if{html}{\out{<SNPdata>}} object. When the value
of this argument is a file, the column must be named as \code{sample}.}
}
\value{
A \if{html}{\out{<SNPdata>}} object where the specified samples have been removed
}
\description{
Remove a set of samples from a \if{html}{\out{<SNPdata>}} object
}
\examples{
\dontrun{
 # create the SNPdata object
 snpdata <- get_snpdata(
   vcf_file = system.file("extdata", "test_data.vcf.gz", package = "mpbr"),
   meta_file = system.file("extdata", "sample_metadata.RDS",
                            package = "mpbr"),
   output_dir = getwd(),
   gaf = file.path("data", "PlasmoDB-68_Pfalciparum3D7_Curated_GO.gaf.gz"),
   gff = file.path("data", "PlasmoDB-68_Pfalciparum3D7.gff")
 )
 # remove samples
 snpdata <- remove_samples(
   snpdata = snpdata,
   samples = c("PA0007-C", "PA0008-C")
 )
} 
}
