% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drop_snps.R
\name{remove_snps}
\alias{remove_snps}
\title{Remove a set of SNPs from a \if{html}{\out{<SNPdata>}} object}
\usage{
remove_snps(snpdata, snps = NULL, chrom = NULL, start = NULL, end = NULL)
}
\arguments{
\item{snpdata}{The input \if{html}{\out{<SNPdata>}} object}

\item{snps}{A data frame with 2 columns named as \code{Chrom} and \code{Pos}}

\item{chrom}{A vector of chromosome names containing the SNPs to be dropped.
This is only used for filtering SNPs within specific regions of the
genome.}

\item{start}{A numeric vector of region's start positions of the region
covering the SNPs to be discarded.}

\item{end}{A numeric vector of region's end positions of the region covering
the SNPs to be discarded.}
}
\value{
A \if{html}{\out{<SNPdata>}} object where the specified SNPs have been removed.
}
\description{
Remove a set of SNPs from a \if{html}{\out{<SNPdata>}} object
}
\details{
When the value for the \code{snps} argument is not NULL, then the rest of
the arguments will be ignored.
}
\examples{
\dontrun{
 # create the SNPdata object
 snpdata <- get_snpdata(
   vcf_file = system.file("extdata", "test_data.vcf.gz", package = "mpbr"),
   meta_file = system.file("extdata", "sample_metadata.RDS",
                            package = "mpbr"),
   output_dir = getwd(),
   gaf = file.path("data", "PlasmoDB-68_Pfalciparum3D7_Curated_GO.gaf.gz"),
   gff = file.path("data", "PlasmoDB-68_Pfalciparum3D7.gff")
 )
 # remove SNPs
 idx_snps <- sample(1:nrow(snpdata[["details"]]), size = 5, replace = FALSE)
 snpdata <- remove_snps(
   snpdata = snpdata,
   snps = snpdata[["details"]][idx_snps, c("Chrom", "Pos")]
 )
}
}
