% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{remove_snps_from_vcf}
\alias{remove_snps_from_vcf}
\title{Delete a set of SNPs from a VCF file}
\usage{
remove_snps_from_vcf(vcf, loci_to_be_retained, path, index = 1L)
}
\arguments{
\item{vcf}{the input VCF file}

\item{loci_to_be_retained}{the name of the file that contains the set of loci
to be retained.}

\item{path}{the path to the directory that contains the input files. This is
also where the output files will be stored.}

\item{index}{the index of the input VCF file}
}
\value{
the path to the filtered VCF file
}
\description{
Delete a set of SNPs from a VCF file
}
\keyword{internal}
