% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_manip.R
\name{rm.prf1}
\alias{rm.prf1}
\title{remove the "exon_" prefix from character string}
\usage{
rm.prf1(x)
}
\arguments{
\item{x}{the input character string}
}
\value{
the input character string without the prefix "exon_"
}
\description{
remove the "exon_" prefix from character string
}
\keyword{internal}
