% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_manip.R
\name{rm.prf2}
\alias{rm.prf2}
\title{remove the "utr_" preffix from a character}
\usage{
rm.prf2(x)
}
\arguments{
\item{x}{the input character string}
}
\value{
the input character string without the preffix "utr_"
}
\description{
remove the "utr_" preffix from a character
}
\keyword{internal}
