% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_manip.R
\name{rm_suf}
\alias{rm_suf}
\title{split a character on '.'}
\usage{
rm_suf(x)
}
\arguments{
\item{x}{the input character string}
}
\value{
a \code{vector} of \code{character} obtained after spliting the input character
based on '.'
}
\description{
split a character on '.'
}
\keyword{internal}
