% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_chrom.R
\name{select_chrom}
\alias{select_chrom}
\title{Select data from specified chromosomes}
\usage{
select_chrom(snpdata, chrom = "all")
}
\arguments{
\item{snpdata}{a \code{SNPdata} object}

\item{chrom}{a vector of chromosome names}
}
\value{
a \code{SNPdata} object with only the data from the specified chromosomes
}
\description{
Select data from specified chromosomes
}
