% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pop_genetics.R
\name{simulate_ys_hmm}
\alias{simulate_ys_hmm}
\title{Title}
\usage{
simulate_ys_hmm(freqquencies, distances, k, r, epsilon)
}
\arguments{
\item{freqquencies}{}

\item{distances}{}

\item{k}{}

\item{r}{}

\item{epsilon}{}
}
\description{
Title
}
