% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_snpdata.R
\name{validate_snpdata}
\alias{validate_snpdata}
\title{Validate SNPdata object}
\usage{
validate_snpdata(x)
}
\arguments{
\item{x}{An object of class SNPdata}
}
\value{
Invisibly returns an object of class SNPdata
}
\description{
Validate SNPdata object
}
\keyword{internal}
