% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_metadata.R
\name{add_metadata}
\alias{add_metadata}
\title{Build the sample metadata table}
\usage{
add_metadata(sample_ids, metadata, output_dir)
}
\arguments{
\item{sample_ids}{a \code{vector} of sample IDs. This should be in the same order
as they appear in the input VCF file.}

\item{metadata}{the path to the sample metadata file}

\item{output_dir}{the path to the output directory}
}
\value{
an object of class \code{data.frame} that contains the sample metadata
}
\description{
Build the sample metadata table
}
\details{
In addition to generating the sample metadata file, the function will also
create a file, named as \code{samples_in_metadata_not_in_vcf.txt}, if there are
samples in the metadata file that were not found in the VCF file.
}
\keyword{internal}
