% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_manip.R
\name{add_metadata}
\alias{add_metadata}
\title{Build the sample metadata table}
\usage{
add_metadata(sample_ids, metadata)
}
\arguments{
\item{sample_ids}{a \code{vector} of sample IDs. This should be in the same order
as they appear in the input VCF file.}

\item{metadata}{the path to the sample metadata file}
}
\value{
an object of class \code{data.frame} that contains the sample metadata
}
\description{
Build the sample metadata table
}
\keyword{internal}
