% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_fws.R
\name{calculate_fws}
\alias{calculate_fws}
\title{Calculate the complexity of the infection (COI) in every sample}
\usage{
calculate_fws(snpdata, threshold = 0.95)
}
\arguments{
\item{snpdata}{a \code{SNPdata} object}

\item{threshold}{a \code{numeric} value between 0 and 1. This will be used to
categorize the infections into monoclonal (Fws > threshold) or polyclonal
(Fws <= threshold)}
}
\value{
a \code{SNPdata} object with 2 additional columns in the meta table
\enumerate{
\item Fws: within host genetic diversity value
\item COI: the complexity of infection: 1 for Fws>0.95, 2 for Fws<=0.95
}
}
\description{
The COI is estimated here through the within host genetic diversity (Fws).
using the {moimix} R package.
}
\examples{
\dontrun{
  snpdata <- calculate_fws(snpdata, threshold = 0.95)
 }

}
