% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_manip.R
\name{calculate_Fws}
\alias{calculate_Fws}
\title{Calculate the complexity of the infection (COI) in every sample}
\usage{
calculate_Fws(snpdata)
}
\arguments{
\item{snpdata}{a \code{SNPdata} object}
}
\value{
a \code{SNPdata} object with 2 additional columns in the meta table
\enumerate{
\item Fws: within host genetic diversity value
\item COI: the complexity of infection: 1 for Fws>0.95, 2 for Fws<=0.95
}
}
\description{
The COI is estimated here through the within host genetic diversity (Fws).
using the {moimix} R package.
}
\examples{
\dontrun{
  snpdata <- calculate_Fws(snpdata)
 }
 
}
