% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pop_genetics.R
\name{calculate_IBS}
\alias{calculate_IBS}
\title{Generate dissimilarity matrix (1-IBS) between all pairs of isolates}
\usage{
snpdata = calculate_IBS(snpdata, mat.name="GT")
}
\arguments{
\item{snpdata}{SNPdata object}

\item{mat.name}{the name of the genotype table to be used. default="GT"}
}
\value{
SNPdata object with an extra field: IBS
}
\description{
Generate dissimilarity matrix (1-IBS) between all pairs of isolates
}
