% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pop_genetics.R
\name{calculate_iR}
\alias{calculate_iR}
\title{Calculate iR index to detect loci with excess of IBD sharing}
\usage{
snpdata = calculate_iR(snpdata, mat.name="Phased", family="Location", number.cores=4)
}
\arguments{
\item{snpdata}{SNPdata object}

\item{mat.name}{the name of the genotype table to be used. default="Phased"}

\item{family}{the name of the column, in the metadata table, to be used to represent the sample's population}

\item{number.cores}{the number of cores to be used. default=4}
}
\value{
SNPdata object with an extra field: iR
}
\description{
Calculate iR index to detect loci with excess of IBD sharing
}
