% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pop_genetics.R
\name{calculate_ld}
\alias{calculate_ld}
\title{Calculate LD R^2 between pairs of loci}
\usage{
calculate_ld(snpdata, min_r2 = 0.2, inter_chrom = FALSE, chroms = NULL)
}
\arguments{
\item{snpdata}{a \code{SNPdata} object}

\item{min_r2}{the minimum r2 value below which the LD value is not reported}

\item{inter_chrom}{a \code{logical} that determines whether to calculate the
inter-chromosomal LD or not. Default is \code{FALSE}.}

\item{chroms}{a character vector of chromosome names. If provided, LD will be
calculated only across these chromosomes.}
}
\value{
a \code{SNPdata} object with an extra field named as \strong{LD}
}
\description{
Calculate LD R^2 between pairs of loci
}
\details{
The output file from LD calculation could be large. In order to
reduce the sizze of that file, we recommend to specify the list of
chromosomes for which LD should be calculated using the \code{chroms} option.
}
\examples{
\dontrun{
  snpdata <- calculate_ld(snpdata, min_r2=0.2, inter_chrom=FALSE,
    chroms=c("Pf3D7_04_v3","Pf3D7_05_v3"))
 }
}
