% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pop_genetics.R
\name{calculate_wcFst}
\alias{calculate_wcFst}
\title{Calculate Weir & Cockerham's Fst}
\usage{
snpdata = calculate_wcFst(snpdata, groups=c("Senegal","Gambia"), from="Country")
}
\arguments{
\item{snpdata}{SNPdata object}

\item{from}{the metadata column from which these groups belong to}

\item{groups}{a vector of string. the differentiation will be estimated between these groups of samples}
}
\value{
SNPdata object with an extra field: Fst. This is a list of data frames that contain the result from a specific comparison. Every data frame contains N rows (where N is the number of loci) and the following 7 columns:
\enumerate{
\item the chromosome ID
\item the SNPs positions
\item the allele frequency in the first group
\item the allele frequency in the second group
\item the resulting Fst values
\item the p-values associated with the Fst results
\item the p-values corrected for multiple testing using the Benjamini-Hochberg method
}
}
\description{
Calculate Weir & Cockerham's Fst
}
