% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_manip.R
\name{compute_MAF}
\alias{compute_MAF}
\title{Calculate minor allele frequency (MAF)}
\usage{
compute_MAF(snpdata, include_het = FALSE, mat_name = "GT")
}
\arguments{
\item{snpdata}{a \code{SNPdata} object}

\item{include_het}{whether to account for the heterozygous allele or not.
This is only used when \code{mat_name = "GT"}}

\item{mat_name}{the name of the matrix to use. default is "GT"}
}
\value{
a \code{SNPdata} object with 2 additional columns in the \strong{details}
table.
\enumerate{
\item MAF: minor allele frequency of each SNPs
\item MAF_allele: 1 if the alternate allele is the minor allele.
0 otherwise
}
}
\description{
Uses the \code{SNPdata} object to calculate the MAF at every loci
}
\details{
if \code{include_het = FALSE}, the mixed allele will not be considered in
the MAF calculation
}
\examples{
\dontrun{
  snpdata <- compute_MAF(
   snpdata,
   include_het = FALSE,
   mat_name = "GT"
 )
 }
}
