% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pop_genetics.R
\name{compute_rhat_hmm}
\alias{compute_rhat_hmm}
\title{Title}
\usage{
compute_rhat_hmm(freqquencies, distances, ys, epsilon)
}
\arguments{
\item{freqquencies}{the freq}

\item{distances}{the dist}

\item{ys}{param1}

\item{epsilon}{param2}
}
\value{
a matrix
}
\description{
Title
}
\keyword{internal}
