% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pop_genetics.R
\name{construct_genotype_file}
\alias{construct_genotype_file}
\title{construct genotype file}
\usage{
construct_genotype_file(pops, details, mat, metadata, from, sweep_regions)
}
\arguments{
\item{pops}{pops}

\item{details}{details}

\item{mat}{mat}

\item{metadata}{metadata}

\item{from}{from}

\item{sweep_regions}{sweep_regions}
}
\value{
test
}
\description{
construct genotype file
}
