% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pop_genetics.R
\name{create_relatedness_matrix}
\alias{create_relatedness_matrix}
\title{create relatedness matrix}
\usage{
create_relatedness_matrix(ibd, metadata, pops, from)
}
\arguments{
\item{ibd}{ibd}

\item{metadata}{metadata}

\item{pops}{pops}

\item{from}{from}
}
\value{
a test
}
\description{
create relatedness matrix
}
