% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drop_samples.R
\name{drop_samples}
\alias{drop_samples}
\title{Drop a specific set of samples from the \code{SNPdata} object}
\usage{
drop_samples(snpdata, samples_to_be_dropped)
}
\arguments{
\item{snpdata}{a \code{SNPdata} object}

\item{samples_to_be_dropped}{a vector of samples to be dropped}
}
\value{
a \code{SNPdata} object where the specified samples have been removed
}
\description{
Drop a specific set of samples from the \code{SNPdata} object
}
\examples{
\dontrun{
  snpdata <- drop_samples(snpdata, samples_to_be_dropped)
 }

}
