% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_manip.R
\name{drop_snps}
\alias{drop_snps}
\title{Drop set a SNPs}
\usage{
snpdata = drop_snps(snpdata, snp.to.be.dropped=NA, chrom="Pf3D7_05_v3", start=100, end=500)
}
\arguments{
\item{snpdata}{a SNPdata object}

\item{snp.to.be.dropped}{a data frame with 2 columns "Chrom" and "Pos"}

\item{chrom}{the chromosome from which loci should be dropped}

\item{start}{the starting position of the region to be discarded}

\item{end}{the end position of the region to be discarded}
}
\value{
a SNPdata object where the specified SNPs have been removed
}
\description{
remove a set of SNPs from the SNPdata object
}
\details{
when snp.to.be.dropped is not set to NA (i.e. the genomic coordinates of snps to be removed are in a data frame), then the rest of the arguments can be ignored or set to NA (chrom=NA, start=NA, end=NA)
}
