% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drop_snps.R
\name{drop_snps_using_gc}
\alias{drop_snps_using_gc}
\title{Drops SNPs from a \code{SNPdata} object using their genomic coordinates}
\usage{
drop_snps_using_gc(snpdata, snp_to_be_dropped)
}
\description{
Drops SNPs from a \code{SNPdata} object using their genomic coordinates
}
\keyword{internal}
