% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access_data_in_vcf_file.R
\name{extract_genotype}
\alias{extract_genotype}
\title{Extract either the genotypes (GT) or the allelic depth (AD) from a variant
call file (VCF).}
\usage{
extract_genotype(vcf_data, which = "GT", nthreads)
}
\arguments{
\item{vcf_data}{A \if{html}{\out{<data table>}} object obtained after reading the input VCF
file.}

\item{which}{A character that denotes the data of interest. The possible
values are:
\enumerate{
\item GT: to extract the genotype data (FORMAT/GT)
\item AD: to extract the allelic depth (FORMAT/AD)
}}

\item{nthreads}{A numeric that represents the number of threads to use during
data import and processing}
}
\value{
A data frame with the genotype data of interest, together with the
SNPs genomic coordinates and their associated alleles (if which = "GT").
}
\description{
Extract either the genotypes (GT) or the allelic depth (AD) from a variant
call file (VCF).
}
\examples{
\dontrun{
  # read in the VCF file
  vcf_data <- data.table::fread(
    file = system.file("extdata", "test_data.vcf.gz", package = "mpbr"),
    skip = "#CHROM",
    header = TRUE,
    sep = "\t",
    nThread = nthreads
  )
  # extract genotype matrix
  genotypes <- extract_genotype(
    vcf_data = vcf_data,
    which = "GT",
    nthreads = 5
  )
  # extract allelic depth matrix
  genotypes <- extract_genotype(
    vcf_data = vcf_data,
    which = "AD",
    nthreads = 5
  )
}
}
\author{
Banky
}
