% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_snps_and_samples.R
\name{filter_snps_samples}
\alias{filter_snps_samples}
\title{Filter loci and samples (requires \strong{bcftools} and \strong{tabix} to be installed)}
\usage{
filter_snps_samples(
  snpdata,
  soft = FALSE,
  min_qual = 10L,
  max_missing_sites = 0.2,
  max_missing_samples = 0.2,
  maf_cutoff = 0.01
)
}
\arguments{
\item{snpdata}{An object of class \code{SNPdata}}

\item{min_qual}{the minimum call quality score below which a loci will be
discarded. default = 10}

\item{max_missing_sites}{the maximum fraction of missing sites above which
a sample should be discarded. default = 0.2}

\item{max_missing_samples}{the maximum fraction of missing samples above
which a loci should be discarded. default = 0.2
@param soft A boolean that specify whether to perform iterative filtering
or not. Iterative filtering consists in soft filtering the input object by
varying the missingness cut-off from 0.8 to 0.1, and the MAF cut-off from
0.4 to 0.1. Default is \code{FALSE}.}

\item{maf_cutoff}{the MAF cut-off. loci with a MAF < maf_cutoff will be
discarded}
}
\value{
a filtered SNPdata object
}
\description{
This function filters the SNPs and samples based on the specified conditions
}
\examples{
\dontrun{
 snpdata <- filter_snps_samples(
  snpdata,
  soft                = FALSE,
  min_qual            = 10,
  max_missing_sites   = 0.2,
  max_missing_samples = 0.2,
  maf_cutoff          = 0.01
 )
 }
}
