% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_snps_and_samples.R
\name{filter_soft}
\alias{filter_soft}
\title{Soft filter on SNPs and samples}
\usage{
filter_soft(details, meta)
}
\arguments{
\item{details}{The \code{SNPdata} object \code{details} data frame.}

\item{meta}{The \code{SNPdata} object \code{meta} data frame.}
}
\value{
Displays the number of SNPs and samples that will be left after
applying the different cut-offs (from 0.8 to 0.1 for missingness, and
0.4 to 0.1 for MAF).
}
\description{
Soft filter on SNPs and samples
}
\keyword{internal}
