% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_manip.R
\name{gene_annotation}
\alias{gene_annotation}
\title{Associate each SNPs to a gene name on which it belongs}
\usage{
gene_annotation(target_gtf, genomic_coordinates)
}
\arguments{
\item{target_gtf}{a \code{data.frame} with the gene annotation}

\item{genomic_coordinates}{a \code{data.frame} with the SNPs genomic coordinates}
}
\value{
a \code{vector} with gene annotation. This should be of the same length as
the number of SNPs.
}
\description{
Associate each SNPs to a gene name on which it belongs
}
\keyword{internal}
