% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation.R
\name{get_clean_name}
\alias{get_clean_name}
\title{clean gene names}
\usage{
get_clean_name(y)
}
\arguments{
\item{y}{the gene name}
}
\value{
a string with the cleaned gene name
}
\description{
clean gene names
}
\keyword{internal}
