% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation.R
\name{get_gene_annotation}
\alias{get_gene_annotation}
\title{Add gene ontology and names annotation details to every SNPs in the table
that contains their genomic coordinates}
\usage{
get_gene_annotation(genomic_coordinates, go, bed, num_cores = 4L)
}
\arguments{
\item{genomic_coordinates}{the table with SNPs genomic coordinates}

\item{go}{a \code{data.frame} with the gene ontology annotation details}

\item{bed}{a \code{data.frame} with the gene name annotation}

\item{num_cores}{the number of cores to be used}
}
\value{
an object of type \code{data.frame} with the SNPs genomic coordinates and
their corresponding annotation details
}
\description{
Add gene ontology and names annotation details to every SNPs in the table
that contains their genomic coordinates
}
\keyword{internal}
