% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pop_genetics.R
\name{get_genotypes}
\alias{get_genotypes}
\title{Title}
\usage{
get_genotypes(
  ped_map,
  reference_ped_map = NULL,
  maf = 0.01,
  isolate_max_missing = 0.1,
  snp_max_missing = 0.1,
  chromosomes = NULL,
  input_map_distance = "cM",
  reference_map_distance = "cM"
)
}
\arguments{
\item{ped_map}{a \code{list} of 2 elements of type data frame with the data in ped
and map format}

\item{reference_ped_map}{a \code{list} of 2 elements of type data frame with the
reference data in ped and map format}

\item{maf}{A numeric value denoting the smallest minor allele frequency
allowed in the analysis. The default value is 0.01}

\item{isolate_max_missing}{A numeric value denoting the maximum proportion of
missing data allowed for each isolate. The default value is 0.1}

\item{snp_max_missing}{A numeric value denoting the maximum proportion of
missing data allowed for each SNP. The default value is 0.1}

\item{chromosomes}{A vector containing a subset of chromosomes to perform
formatting on. The default value is chromosomes=NULL which will reformat
all genotypes for all chromosomes in the MAP data frame}

\item{input_map_distance}{A character string of either "M" or "cM" denoting
whether the genetic map distances in the input MAP data frame are in
Morgans (M) or centi-Morgans (cM). The default is cM}

\item{reference_map_distance}{A character string of either "M" or "cM"
denoting whether the genetic map distances in the reference MAP data frame
are in Morgans (M) or centi-Morgans (cM). The default is cM}
}
\value{
A \code{list} of two objects named pedigree and genotypes
}
\description{
Title
}
