% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_snpdata.R
\name{get_snpdata}
\alias{get_snpdata}
\title{Create A \code{SNPdata} object}
\usage{
get_snpdata(vcf_file, meta_file, output_dir, gaf, gff, num_threads = 4L)
}
\arguments{
\item{vcf_file}{A character with the path to the input VCF file (required)}

\item{meta_file}{A character with the path to the sample metadata file
(required)}

\item{output_dir}{A character with the path to the folder where the output
and temporary files will be stored (optional)}

\item{gaf}{A character with the path to the file with the gene ontology
annotation file (required). This can be downloaded from
\href{https://plasmodb.org/plasmo/app/downloads/Current_Release/Pfalciparum3D7/gaf/}{PlasmoDB} # nolint: line_length_linter}

\item{gff}{A character with the path to the gene annotation file. This file
can be downloaded from
\href{https://plasmodb.org/plasmo/app/downloads/Current_Release/Pfalciparum3D7/gff/}{PlasmoDB} # nolint: line_length_linter}

\item{num_threads}{The number of threads to be during data processing}
}
\value{
An object of class \if{html}{\out{<SNPdata>}} that contains the following 4 elements:
\enumerate{
\item meta: A \code{data.frame} that with the sample's metadata
\item details: A \code{data.frame} with the SNPs genomic coordinates, and their
associated annotations from both the VCF and input annotation files
\item GT: An integer \code{matrix} with the sample's genotype data where:
0 represents the reference allele, 1 the alternative allele, 2 is a
mixed allele, and NA a missing genotype.
\item vcf: the full path to the input VCF file from which the data is
generated.
}
}
\description{
The function generates the input object type require to use functions in
the {mpbr} package.
}
\examples{
\dontrun{
  snpdata <- get_snpdata(
    vcf_file = system.file("extdata", "test_data.vcf.gz", package = "mpbr"),
    meta_file = system.file("extdata", "sample_metadata.RDS",
                            package = "mpbr"),
    output_dir = tempdir(),
    gaf = file.path("data", "PlasmoDB-68_Pfalciparum3D7_Curated_GO.gaf.gz"),
    gff = file.path("data", "PlasmoDB-68_Pfalciparum3D7.gff")
  )
}
}
