% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_manip.R
\name{impute}
\alias{impute}
\title{Impute missing genotypes for 1 SNP across all samples}
\usage{
impute(genotype)
}
\arguments{
\item{genotype}{}
}
\value{
the input vector where all missing alleles have been imputed.
}
\description{
Impute missing genotypes for 1 SNP across all samples
}
\keyword{internal}
