% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute_missing_genotypes.R
\name{impute_one_allele}
\alias{impute_one_allele}
\title{Impute missing genotypes for one SNP across all samples}
\usage{
impute_one_allele(genotype)
}
\arguments{
\item{genotype}{A vector of integers representing the alleles across all
samples at a given locus.}
}
\value{
The input vector where all missing alleles have been imputed.
}
\description{
Impute missing genotypes for one SNP across all samples
}
\keyword{internal}
