% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pop_genetics.R
\name{loglikelihood}
\alias{loglikelihood}
\title{Title}
\usage{
loglikelihood(k, r, ys, f, gendist, epsilon, rho = 7.4 * 10^(-7))
}
\arguments{
\item{k}{}

\item{r}{}

\item{ys}{}

\item{f}{}

\item{gendist}{}

\item{epsilon}{}

\item{rho}{}
}
\description{
Title
}
