% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_manip.R
\name{phase_mixed_genotypes}
\alias{phase_mixed_genotypes}
\title{Phase mixed genotypes}
\usage{
phase_mixed_genotypes(snpdata, nsim = 100L)
}
\arguments{
\item{snpdata}{a \code{SNPdata} object}

\item{nsim}{an integer that represents the number of simulations to be
performed}
}
\value{
a \code{SNPdata} object with an additional table named as \strong{Phased}. This
object will contain the phased genotype data
}
\description{
Mixed genotype phasing is performed based on the number of reads supporting
each allele at an heterozygous site. The phasing will be run \code{nsim} times and
phased data with the highest correlation between MAF from raw data and MAF
from phased data will be retained.
}
\details{
When both alleles are not supported by any read or the total number
of reads supporting both alleles at a given site is < 5, the genotype will
be phased based on a Bernoulli distribution using the MAF as a parameter.
Similarly, when the total number of reads is > 5 and the number of reads
supporting one of the allele is not 2 times the number of the other,
the genotype is phased using a Bernoulli distribution.
}
\examples{
\dontrun{
  snpdata <- phase_mixed_genotypes(snpdata)
 }

}
