% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_manip.R
\name{print.SNPdata}
\alias{print.SNPdata}
\title{Print the \code{SNPdata} object}
\usage{
\method{print}{SNPdata}(snpdata)
}
\arguments{
\item{snpdata}{the \code{SNPdata} object}
}
\description{
Print the \code{SNPdata} object
}
\examples{
\dontrun{
  print(snpdata)
}
  
}
